Rewrite the solving algorithm
The algorithm was rewritten to do the following:

* Get the list of elements with the lowest amount of valid values
  available
* Pick an element randomly from the list
* Pick a valid value randomly from the list of available values
* If no valid values were available, then backtrack
  - If puzzle is less than 50% solved, go to the earliest change with
    another value available
  - Otherwise, just go back to the most recent change with another value
    available.

The backtracking change seems to generally avoid the worst case types of scenarios
which make this algorithm O(n!) which is having to backtrack for
every change with multiple values to solve the puzzle.

This rewrite also  allows the class to be used to create new puzzles.
This can be accomlished by solving an empty board, then removing
elements randomly to create the desired difficultly of puzzle.