Fix stack underflow issue in solve method.
An unsolvable puzzle layout would cause the stack to underflow.

Added check for stack underflow and return Unsolvable puzzle layout in
that situation.

Restored to just popping the stack to backtrack as the
previous bug which wasn't properly setting the elements array was the
reason for extremely long runs. Additionally, moving forward through the
stack made it more difficult to determine an unsolvable layout
condition.

Bumped the attemptCount check to 5000. A few valid, but very challenging
puzzles can trigger the O(n!) worse case, but most are solved within
that limit.